perf: use already calculated total width and height
render() method calls getTotalWidth and getTotalHeight methods and saves them to variables named totalWidth and totalHeight.

However, render method also calls isOverOverlay method, which also calls and calculates getTotalWidth and getTotalHeight again.

Added a default arguments to isOverOverlay method that accepts width and height that calls getTotalWidth and getTotalHeight by default. This makes other calls to the function calculate the width same as before, while the call from render is replaced to pass the already calculated totalWidth and totalHeight values.

This is improves both consistency (the totalWidth and totalHeight used in render and isOverOverlay will be guaranteed to be exactly the same in all cases) and performance due to no recalculation of the width and height (the getTotalWidth and getTotalWeight methods loop over all lines and perform other method calls, so they are not the cheapest methods to be called multiple times).