perf: cache tab list parsing per tick
Implements a per tick cache for the tab list information.

Does not seem to cause any issues during my testing and improves performance.

Note that the cache list has to be immutable and replaced each tick, because the method is called from both netty and render threads so it has to be thread safe. When testing with a mutable list and doing clear + add on updateCache each tick, there was ConcurrentModificationException's.

This approx. a 5% optimization inside the netty thread as the mod seems to call it 2 times inside there (one inside BurrowDetector#onParticleReceive and another in PreciseGuessBurrow#onReceiveParticle). I haven't seen any significiant difference in Render thread time spent.