fix: re-entrancy of Register with TabList cache
When testing merging Arrow-Guess branch on top of Diana-V2 branch, the game crashes on startup with a ConcurrentModificationException:

---- Minecraft Crash Report ----
// Quite honestly, I wouldn't worry myself about that.

Time: 2025-12-13 18:10:12
Description: Unexpected error

java.util.ConcurrentModificationException
	at java.base/java.util.ArrayList$Itr.checkForComodification(ArrayList.java:1096)
	at java.base/java.util.ArrayList$Itr.next(ArrayList.java:1050)
	at knot//net.sbo.mod.utils.events.TickScheduler._init_$lambda$1(TickScheduler.kt:12)
	at knot//net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents.lambda$static$2(ClientTickEvents.java:43)
	at knot//net.minecraft.class_310.handler$boc000$fabric-lifecycle-events-v1$onEndTick(class_310.java:9092)
	at knot//net.minecraft.class_310.method_1574(class_310.java:1960)
	at knot//net.minecraft.class_310.method_1523(class_310.java:1324)
	at knot//net.minecraft.class_310.method_1514(class_310.java:947)
	at knot//net.minecraft.client.main.Main.ixeris$runRenderThread(Main.java:1081)
	at knot//net.minecraft.client.main.Main.md1fe659$ixeris$lambda$ixeris$main$0$0(Main.java:1050)
	at java.base/java.lang.Thread.run(Thread.java:1474)

The likely trigger is the 16268aa commit, which calls World.getWorld() inside a Register.onTick block. World.getWorld() calls TabList.findInfo, and that accesses the lazy initOnce variable, which does updateCache and calls Register.onTick to schedule future updateCache calls. Trying to Register a tick task while inside already a Registered tick task fails with CME as the tasks list is modified while it's being iterated.

This likely would not happen without the TabList cache, so the cache is at fault - but the trigger is a new commit in Arrow-Guess branch as I didn't have this crash before with the cache. There might be more places that call World.getWorld() inside a tick task, but they probably did not cause an issue because it was called earlier from a non-tick task, as initOnce only calls Register.onTick on first access.

The Register class could be changed to support registering tick tasks when inside one, or the TabList class can be made to use END_CLIENT_TICK directly, but I figured the simplest fix was to always init the tab list updateCache task at a predictable time which is not inside a re-entrant tick task (at mod init). This successfully fixes the issue for now and no further refactor is needed. I've placed the TabList.init call before all other inits just in case.