fix: correct mcmeta for sound resourcepack
Fixed currentResourcePackFormat() returning 48 (catch block) on 1.21.10 because the Int cast was throwing an exception, since the method returns a PackVersion record object, not anyting that can be converted to an Int automatically. This fix uses the major() method from PackVersion instead of trying to cast PackVersion to an Int.

Mojang replaced supported_formats with min_format and max_format in 1.21.9's first snapshot and kept it that way in final release and 1.21.10 as well. This commit in addition to fixing currentResourcePackFormat also migrates to use the new fields, making the resourcepack show as compatible in the resourcepack selection screen.