add: money hour that always shows before hover
Better feeling to see it update live while you are playing instead of stopping to type !profit or hovering over the element. Hovering over in fact does not work for me but that might just be something on 1.21.10 or on my side.

Mainly just copied the burrows/hr code and made it work on string instead of number. Looks exactly same as the burrows/hr one.

Let me know if any changes are needed!